add float

+++ pytorch_lightning/metrics/functional/iou.py
@@ -29,7 +29,7 @@ def _iou_from_confmat(
     union = confmat.sum(0) + confmat.sum(1) - intersection
 
     # If this class is absent in both target AND pred (union == 0), then use the absent_score for this class.
-    scores = intersection / union
+    scores = intersection.float() / union.float()
     scores[union == 0] = absent_score
 
     # Remove the ignored class index from the scores.
